episode-projekter/juleplayliste/pyproject.toml
[tool.poetry]
name = "juleplayliste"
version = "0.1.0"
description = ""
authors = ["Henrik Gregersen <support@allekankode.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
spotipy = "^2.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
